# apps/api/internal/docs/openapi.yaml
openapi: 3.1.0
info:
  title: SaveSphere API
  description: |
    Personal finance and budget tracking API. Authentication uses httpOnly cookies
    (access_token + refresh_token). Pass `credentials: include` on all requests.
  version: 1.0.0
  contact:
    name: SaveSphere
servers:
  - url: http://localhost:3000
    description: Local development
paths:
  /api/health:
    get:
      tags: [System]
      summary: Health check
      operationId: healthCheck
      responses:
        "200":
          description: API is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a new account
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Account already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/auth/login:
    post:
      tags: [Auth]
      summary: Sign in
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          headers:
            Set-Cookie:
              description: access_token + refresh_token httpOnly cookies
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Email not verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Account temporarily locked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/auth/logout:
    post:
      tags: [Auth]
      summary: Sign out
      operationId: logout
      responses:
        "200":
          description: Logged out successfully
          headers:
            Set-Cookie:
              description: Clears access_token and refresh_token cookies
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      operationId: refreshToken
      description: Reads refresh_token from httpOnly cookie and issues new tokens
      responses:
        "200":
          description: Token refreshed
          headers:
            Set-Cookie:
              description: New access_token + refresh_token cookies
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      token:
                        type: string
        "401":
          description: Invalid or expired session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/auth/me:
    get:
      tags: [Auth]
      summary: Get current user
      operationId: getMe
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request password reset
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
      responses:
        "200":
          description: Reset link sent (if account exists)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /api/auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset password with token
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Password updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/auth/verify-email:
    get:
      tags: [Auth]
      summary: Verify email address
      operationId: verifyEmail
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Email verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/auth/google:
    post:
      tags: [Auth]
      summary: Begin Google OAuth flow
      operationId: beginGoogleAuth
      responses:
        "200":
          description: Returns Google authorization URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      url:
                        type: string
                        format: uri
        "503":
          description: Google Sign-In not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/auth/google/callback:
    post:
      tags: [Auth]
      summary: Google OAuth callback
      operationId: googleCallback
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Login successful
          headers:
            Set-Cookie:
              description: access_token + refresh_token cookies
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid state or missing code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/wallets:
    get:
      tags: [Wallets]
      summary: List wallets
      operationId: listWallets
      security:
        - cookieAuth: []
      responses:
        "200":
          description: List of user wallets
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Wallet"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags: [Wallets]
      summary: Create wallet
      operationId: createWallet
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletCreateRequest"
      responses:
        "201":
          description: Wallet created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: "#/components/schemas/Wallet"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/wallets/liquidity:
    get:
      tags: [Wallets]
      summary: Get total liquidity
      operationId: getLiquidity
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Total liquidity amount
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      liquidity:
                        type: string
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/transactions:
    post:
      tags: [Transactions]
      summary: Create transaction
      operationId: createTransaction
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionCreateRequest"
      responses:
        "201":
          description: Transaction created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: "#/components/schemas/Transaction"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/wallets/{walletId}/transactions:
    get:
      tags: [Transactions]
      summary: List transactions for a wallet
      operationId: listTransactions
      security:
        - cookieAuth: []
      parameters:
        - name: walletId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: is_fixed
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Transaction"
        "400":
          description: Invalid wallet ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/transactions/{id}:
    delete:
      tags: [Transactions]
      summary: Delete a transaction
      operationId: deleteTransaction
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Transaction deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid transaction ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/intelligence/safe-to-spend:
    get:
      tags: [Intelligence]
      summary: Safe-to-Spend calculation
      operationId: getSafeToSpend
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Safe to spend amount
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      safe_to_spend:
                        type: string
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/intelligence/advice:
    get:
      tags: [Intelligence]
      summary: AI financial advice
      operationId: getAdvice
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Generated financial advice
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      advice:
                        type: string
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: httpOnly cookie set after login/register/oauth

  schemas:
    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: ok
        app:
          type: string
          example: savesphere-api

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          type: object
          nullable: true

    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        error:
          type: string
          nullable: true

    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        fullName:
          type: string
        authProvider:
          type: string
          example: local
        providerId:
          type: string
          nullable: true
        emailVerified:
          type: boolean
        picture:
          type: string
          nullable: true
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    RegisterRequest:
      type: object
      required: [email, password, fullName]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          format: password
        fullName:
          type: string

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password

    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          properties:
            user:
              $ref: "#/components/schemas/User"
            token:
              type: string

    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email

    ResetPasswordRequest:
      type: object
      required: [token, password]
      properties:
        token:
          type: string
        password:
          type: string
          minLength: 8
          format: password

    Wallet:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
          enum: [Cash, Bank, Credit]
        currency:
          type: string
          minLength: 3
          maxLength: 3
        balance:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    WalletCreateRequest:
      type: object
      required: [name, type, currency]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [Cash, Bank, Credit]
        currency:
          type: string
          minLength: 3
          maxLength: 3
          example: USD

    Transaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        wallet_id:
          type: string
          format: uuid
        category_id:
          type: string
          format: uuid
          nullable: true
        amount:
          type: string
        type:
          type: string
          enum: [Income, Expense, Saving]
        is_fixed:
          type: boolean
        description:
          type: string
        date:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time

    TransactionCreateRequest:
      type: object
      required: [wallet_id, amount, type, date]
      properties:
        wallet_id:
          type: string
          format: uuid
        category_id:
          type: string
          format: uuid
          nullable: true
        amount:
          type: string
        type:
          type: string
          enum: [Income, Expense, Saving]
        is_fixed:
          type: boolean
          default: false
        description:
          type: string
        date:
          type: string
          format: date-time